move getFailingPages to util to share func
