+++ packages/story-editor/src/components/checklist/checks/pageBackgroundLowTextContrast/index.js
@@ -16,3 +16,4 @@
 
 export { default as PageBackgroundTextLowContrast } from './component';
 export { pageBackgroundTextLowContrast } from './check';
+export { getFailingPages } from './utils';
