+++ packages/story-editor/src/components/checklist/checks/pageBackgroundLowTextContrast/utils.js
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * Internal dependencies
+ */
+import { pageBackgroundTextLowContrast } from './check';
+
+export const getFailingPages = async (storyPages, pageSize, currentPage) => {
+  const promises = [];
+  (storyPages || []).forEach((page) => {
+    if (currentPage && currentPage.id !== page.id) {
+      return;
+    }
+    const maybeTextContrastResult = pageBackgroundTextLowContrast({
+      ...page,
+      pageSize,
+    });
+    if (maybeTextContrastResult instanceof Promise) {
+      promises.push(
+        maybeTextContrastResult.then((result) => ({ result, page }))
+      );
+    } else {
+      promises.push(maybeTextContrastResult);
+    }
+  });
+  const awaitedResult = await Promise.all(promises);
+  return awaitedResult.filter(({ result }) => result).map(({ page }) => page);
+};
